Rename service modules and remove InsertResult class
